dom/workers/test/test_bug1063538.html

<!--
2   Any copyright is dedicated to the Public Domain.
3   http://creativecommons.org/publicdomain/zero/1.0/
4 -->
<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1063538
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1063538</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1063538">Mozilla Bug 1063538</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

function runTest() {
  var worker = new Worker("bug1063538_worker.js");

  worker.onmessage = function(e) {
    if (e.data.type == 'finish') {
      ok(e.data.progressFired, "Progress was fired.");
      SimpleTest.finish();
    }
  };

  worker.postMessage(true);
}

SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  SpecialPowers.pushPermissions([{'type': 'systemXHR', 'allow': true, 'context': document}], runTest);
});

</script>
</pre>
</body>
</html>
